Trait vault_registry_rpc::VaultRegistryApiServer

pub trait VaultRegistryApiServer<BlockHash, VaultId, Balance, UnsignedFixedPoint, CurrencyId, AccountId>: Sized + Send + Sync + 'static where
    Balance: Codec + MaybeDisplay + MaybeFromStr,
    UnsignedFixedPoint: Codec + MaybeDisplay + MaybeFromStr,
    CurrencyId: Codec,
    AccountId: Codec, {
    fn get_vault_collateral(
        &self, 
        vault_id: VaultId, 
        at: Option<BlockHash>
    ) -> RpcResult<BalanceWrapper<Balance>>;
    fn get_vaults_by_account_id(
        &self, 
        account_id: AccountId, 
        at: Option<BlockHash>
    ) -> RpcResult<Vec<VaultId>>;
    fn get_vault_total_collateral(
        &self, 
        vault_id: VaultId, 
        at: Option<BlockHash>
    ) -> RpcResult<BalanceWrapper<Balance>>;
    fn get_premium_redeem_vaults(
        &self, 
        at: Option<BlockHash>
    ) -> RpcResult<Vec<(VaultId, BalanceWrapper<Balance>)>>;
    fn get_vaults_with_issuable_tokens(
        &self, 
        at: Option<BlockHash>
    ) -> RpcResult<Vec<(VaultId, BalanceWrapper<Balance>)>>;
    fn get_vaults_with_redeemable_tokens(
        &self, 
        at: Option<BlockHash>
    ) -> RpcResult<Vec<(VaultId, BalanceWrapper<Balance>)>>;
    fn get_issuable_tokens_from_vault(
        &self, 
        vault: VaultId, 
        at: Option<BlockHash>
    ) -> RpcResult<BalanceWrapper<Balance>>;
    fn get_collateralization_from_vault(
        &self, 
        vault: VaultId, 
        only_issued: bool, 
        at: Option<BlockHash>
    ) -> RpcResult<UnsignedFixedPoint>;
    fn get_collateralization_from_vault_and_collateral(
        &self, 
        vault: VaultId, 
        collateral: BalanceWrapper<Balance>, 
        only_issued: bool, 
        at: Option<BlockHash>
    ) -> RpcResult<UnsignedFixedPoint>;
    fn get_required_collateral_for_wrapped(
        &self, 
        amount_btc: BalanceWrapper<Balance>, 
        currency_id: CurrencyId, 
        at: Option<BlockHash>
    ) -> RpcResult<BalanceWrapper<Balance>>;
    fn get_required_collateral_for_vault(
        &self, 
        vault_id: VaultId, 
        at: Option<BlockHash>
    ) -> RpcResult<BalanceWrapper<Balance>>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
        VaultId: Send + Sync + 'static + DeserializeOwned + Serialize,
        Balance: Send + Sync + 'static + DeserializeOwned + Serialize + Codec + MaybeDisplay + MaybeFromStr,
        UnsignedFixedPoint: Send + Sync + 'static + Serialize + Codec + MaybeDisplay + MaybeFromStr,
        CurrencyId: Send + Sync + 'static + DeserializeOwned + Codec,
        AccountId: Send + Sync + 'static + DeserializeOwned + Codec,
    { ... }
}

Server trait implementation for the VaultRegistryApi RPC API.

Required Methods

fn get_vault_collateral(
&self,
vault_id: VaultId,
at: Option<BlockHash>
) -> RpcResult<BalanceWrapper<Balance>>

fn get_vaults_by_account_id(
&self,
account_id: AccountId,
at: Option<BlockHash>
) -> RpcResult<Vec<VaultId>>

fn get_vault_total_collateral(
&self,
vault_id: VaultId,
at: Option<BlockHash>
) -> RpcResult<BalanceWrapper<Balance>>

fn get_premium_redeem_vaults(
&self,
at: Option<BlockHash>
) -> RpcResult<Vec<(VaultId, BalanceWrapper<Balance>)>>

fn get_vaults_with_issuable_tokens(
&self,
at: Option<BlockHash>
) -> RpcResult<Vec<(VaultId, BalanceWrapper<Balance>)>>

fn get_vaults_with_redeemable_tokens(
&self,
at: Option<BlockHash>
) -> RpcResult<Vec<(VaultId, BalanceWrapper<Balance>)>>

fn get_issuable_tokens_from_vault(
&self,
vault: VaultId,
at: Option<BlockHash>
) -> RpcResult<BalanceWrapper<Balance>>

fn get_collateralization_from_vault(
&self,
vault: VaultId,
only_issued: bool,
at: Option<BlockHash>
) -> RpcResult<UnsignedFixedPoint>

fn get_collateralization_from_vault_and_collateral(
&self,
vault: VaultId,
collateral: BalanceWrapper<Balance>,
only_issued: bool,
at: Option<BlockHash>
) -> RpcResult<UnsignedFixedPoint>

fn get_required_collateral_for_wrapped(
&self,
amount_btc: BalanceWrapper<Balance>,
currency_id: CurrencyId,
at: Option<BlockHash>
) -> RpcResult<BalanceWrapper<Balance>>

fn get_required_collateral_for_vault(
&self,
vault_id: VaultId,
at: Option<BlockHash>
) -> RpcResult<BalanceWrapper<Balance>>

Provided Methods

fn into_rpc(self) -> RpcModule<Self> where
BlockHash: Send + Sync + 'static + DeserializeOwned,
VaultId: Send + Sync + 'static + DeserializeOwned + Serialize,
Balance: Send + Sync + 'static + DeserializeOwned + Serialize + Codec + MaybeDisplay + MaybeFromStr,
UnsignedFixedPoint: Send + Sync + 'static + Serialize + Codec + MaybeDisplay + MaybeFromStr,
CurrencyId: Send + Sync + 'static + DeserializeOwned + Codec,
AccountId: Send + Sync + 'static + DeserializeOwned + Codec,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.